pyproject.toml
[project]
name = "template-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.11.6",
]


[[tool.mypy.overrides]]
module = ["yaml.*"]
follow_untyped_imports = true